Call ap_pipe workhorses directly.
This change removes calls to the adapters for doProcessCcd, doDiffIm,
and doAssociation, replacing them with calls directly to the
underlying functions (which are assumed to be renamed). The adapters
were motivated by the assumption that ap_verify should see a single
repository whose internal structure is managed by ap_pipe; both
assumptions are now known to be false.

The repository paths defined here should NOT be the responsibility of
pipeline_driver, and will be removed in future commits.