Modify metrics and error handling.
The new code imports metrics on exit or failure instead of after each
pipeline step, allowing us to, for the moment, query the task object
for metadata. This may need to be reworked depending on how parallel
processing is handled.

This commit makes one behavior change, in that it now tries to extract
metrics from the failed pipeline step. If this process raises an
exception, it will suppress any exception raised by the pipeline. This
is an acceptable cost for the time being.