Make metrics handling per-dataID.
This change involves several coupled decisions, including moving
responsibility for Job-level metadata to AutoJob. The main reason this
commit can't be broken up is because Job metadata requires a data ID,
which is only available during or after pipeline processing.
Therefore, removing runApPipe's dependency on Job and moving Job
creation inside the data ID loop needs to happen simultaneously.

I have not added tests for AutoJob, because the class is being
deprecated in favor of MetricsControllerTask.